Split the Search component into more child components
