+++ client/src/components/Search/Filters/Filters.scss
@@ -1,0 +1,3 @@
+.filters {
+    margin-bottom: 0.5rem;
+}
