+++ client/src/components/Search/Search.scss
@@ -4,12 +4,11 @@
 
 .search-results-page {
     height: 100%;
-    padding-top: 1.5rem;
     display: flex;
     > div {
         flex: 33%;
     }
-    .map-results {
+    .map-column {
         display: flex;
         justify-content: flex-end;
         height: 100%;
@@ -20,7 +19,7 @@
     }
 }
 
-.search-results {
+.search-column {
     width: 40rem;
     @include media-breakpoint-up(lg) {
         max-width: none;
@@ -32,7 +31,7 @@
     display: flex;
     flex-direction: column;
     position: relative;
-    padding: 0 2rem 0 6rem;
+    padding: 0 4rem;
     .search-result {
         margin-bottom: 1.5rem;
     }
@@ -46,10 +45,7 @@
     background-color: white;
     z-index: 1;
     margin-bottom: 0.5rem;
-
-    .filters {
-        margin-bottom: 0.5rem;
-    }
+    padding-bottom: 0.5rem;
 
     .search-info {
         display: flex;
